requirements.txt
streamlit
pandas
altair
databricks-sql-connector
networkx
plotly
